test: move tests into packages
