+++ .markdownlint-cli2.mjs
@@ -2,5 +2,5 @@ import markdownlintConfig from './packages/markdownlint-config/index.js';
 
 export default {
 	config: markdownlintConfig.config,
-	ignores: ['node_modules', 'test', '**/CHANGELOG.md'],
+	ignores: ['node_modules', 'test', '**/CHANGELOG.md', 'packages/markdownlint-config/test/fixtures/bad/*'],
 };
