+++ package.json
@@ -11,7 +11,7 @@
 		"lint": "run-s -l lint:*",
 		"lint:markdownlint": "markdownlint-cli2 \"**/*.md\"",
 		"lint:eslint": "eslint",
-		"test": "node --test test/*.test.js",
+		"test": "npm run test --workspaces",
 		"prepare": "husky"
 	},
 	"devDependencies": {
