+++ packages/eslint-config/test/fixtures/bad/javascript.js
