+++ packages/eslint-config/test/fixtures/bad/json.json
