+++ packages/eslint-config/test/fixtures/bad/jsonc.jsonc
