+++ packages/eslint-config/test/fixtures/good/javascript.js
