+++ packages/eslint-config/test/fixtures/good/json.json
