+++ packages/eslint-config/test/fixtures/good/jsonc.jsonc
