+++ packages/markdownlint-config/test/markdownlint.test.js
@@ -1,7 +1,7 @@
 import test from 'node:test';
 import assert from 'node:assert';
-import config from '../packages/stylelint-config/index.js';
+import config from '../index.js';
 
-test('stylelintConfig', async () => {
+test('markdownlintConfig', async () => {
 	assert.ok(typeof config === 'object');
 });
